A power (pow) function
The power function is used to find the power given two numbers that are the base and exponent. The result is the base raised to the power of the exponent.
The power is calculated using the for loop that runs till the value of the exponent. In each pass, the base value is multiplied with result. After the completion of the for loop, the final value of power is stored in the variable result.